# projects/points-web-app/docs/v0.2/points-markets.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Freeism Points–Markets API",
    "version": "0.2.0",
    "description": "Points-owned canonical contract for the twelve APIs consumed by Markets. Better Auth discovery, authorization, token, introspection and revocation endpoints remain standard provider endpoints and are intentionally not duplicated as application operations."
  },
  "servers": [
    {
      "url": "https://points.freeism.app"
    }
  ],
...
  "paths": {
...
    "/api/v1/point-reservations/status": {
      "post": {
        "operationId": "getPointReservationStatus",
        "summary": "Get statuses for owned point reservations",
        "security": [
          {
            "m2mOAuth": ["points.reservations.status"]
          }
        ],
        "parameters": [
          {
            "name": "X-Request-Id",
            "in": "header",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/OpaqueId"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReservationStatusRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "headers": {
              "Cache-Control": {
                "description": "Exact cache policy: private, no-store",
                "required": true,
                "schema": {
                  "type": "string",
                  "const": "private, no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationStatusResponse"
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "default": {
            "description": "RFC 9457 Problem Details",
            "headers": {
              "Cache-Control": {
                "description": "Exact cache policy: private, no-store",
                "required": true,
                "schema": {
                  "type": "string",
                  "const": "private, no-store"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "x-freeism-request-body-max-bytes": 1048576,
        "x-freeism-cache-control": "private, no-store",
        "x-freeism-ownership": {
          "onlyCallingClientOwnedResources": true,
          "unknownOrOtherClientStatus": 404,
          "unknownOrOtherClientCode": "RESOURCE_NOT_FOUND"
        }
      }
    },
...
    "/api/v1/point-reservations/release": {
      "post": {
        "operationId": "releasePointReservation",
        "summary": "Release an active owned reservation",
        "security": [
          {
            "m2mOAuth": ["points.reservations.release"]
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "description": "Opaque idempotency key. The same key and canonical payload replay the initial HTTP status and domain result.",
            "schema": {
              "$ref": "#/components/schemas/OpaqueId"
            }
          },
          {
            "name": "X-Request-Id",
            "in": "header",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/OpaqueId"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReleaseReservationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "headers": {
              "Cache-Control": {
                "description": "Exact cache policy: private, no-store",
                "required": true,
                "schema": {
                  "type": "string",
                  "const": "private, no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReleaseReservationResponse"
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "default": {
            "description": "RFC 9457 Problem Details",
            "headers": {
              "Cache-Control": {
                "description": "Exact cache policy: private, no-store",
                "required": true,
                "schema": {
                  "type": "string",
                  "const": "private, no-store"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "x-freeism-request-body-max-bytes": 1048576,
        "x-freeism-idempotency": {
          "replayHttpStatus": "preserve-initial",
          "replayDomainResult": true,
          "requestIdMayBeRegenerated": true,
          "sameKeyDifferentPayloadStatus": 409,
          "sameKeyDifferentPayloadCode": "IDEMPOTENCY_KEY_REUSED"
        },
        "x-freeism-cache-control": "private, no-store",
        "x-freeism-state-transition": {
          "from": "ACTIVE",
          "to": "RELEASED",
          "capturedReleaseAllowed": false,
          "onlyCallingClientOwnedResource": true
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "userOAuth": {
        "type": "oauth2",
        "description": "JWT Authorization Code + PKCE access tokens for ordinary user operations. Refresh tokens may renew only this user scope set.",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://points.freeism.app/api/auth/oauth2/authorize",
            "tokenUrl": "https://points.freeism.app/api/auth/oauth2/token",
            "refreshUrl": "https://points.freeism.app/api/auth/oauth2/token",
            "scopes": {
              "openid": "OpenID Connect identity",
              "profile": "Basic profile",
              "offline_access": "Refresh token access",
              "points.connection.read": "Read the active Points connection",
              "points.balance.read": "Check point balances",
              "points.reservations.create": "Create point reservations"
            },
            "x-freeism-pkce": {
              "required": true,
              "method": "S256"
            }
          }
        },
        "x-freeism-access-token-format": "jwt",
        "x-freeism-principal-class": "user"
      },
      "connectionUnlinkOAuth": {
        "type": "oauth2",
        "description": "One-time Authorization Code + PKCE token after Google fresh authentication for normal unlink only.",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://points.freeism.app/api/auth/oauth2/authorize",
            "tokenUrl": "https://points.freeism.app/api/auth/oauth2/token",
            "scopes": {
              "points.connection.unlink": "Deactivate exactly one active Points connection"
            },
            "x-freeism-pkce": {
              "required": true,
              "method": "S256"
            }
          }
...
    "responses": {
      "RateLimited": {
        "description": "RFC 9457 rate limit Problem Details",
        "headers": {
          "Cache-Control": {
            "description": "Exact cache policy: private, no-store",
            "required": true,
            "schema": {
              "type": "string",
              "const": "private, no-store"
            }
          },
          "Retry-After": {
            "description": "Non-empty HTTP Retry-After value",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/RateLimitedProblem"
            }
          }
        }
      }
    },
    "schemas": {
      "OpaqueId": {
        "type": "string",
        "minLength": 1,
        "maxLength": 255
      },
      "ReservationKey": {
        "type": "string",
        "minLength": 1,
        "maxLength": 512
      },
      "Sha256Hash": {
        "type": "string",
        "pattern": "^sha256:[0-9a-f]{64}$"
      },
      "UtcInstant": {
        "type": "string",
        "format": "date-time"
      },
...
      "NonNegativeSafeInteger": {
        "type": "integer",
        "minimum": 0,
        "maximum": 9007199254740991
      },
...
      "RequestMeta": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "requestId": {
            "$ref": "#/components/schemas/OpaqueId"
          }
        },
        "required": ["requestId"]
      },
      "ValidationError": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "row": {
            "$ref": "#/components/schemas/NonNegativeSafeInteger"
          },
          "field": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255
          },
          "code": {
            "type": "string",
            "pattern": "^[A-Z][A-Z0-9_]*$"
          },
          "message": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1000
          }
        },
        "required": ["code"]
      },
      "ProblemDetails": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string",
            "minLength": 1
          },
          "status": {
            "type": "integer",
            "minimum": 400,
            "maximum": 599
          },
          "detail": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000
          },
          "instance": {
            "type": "string",
            "format": "uri-reference"
          },
          "code": {
            "type": "string",
            "enum": [
              "MALFORMED_REQUEST",
              "AUTHENTICATION_REQUIRED",
              "INVALID_ACCESS_TOKEN",
              "INSUFFICIENT_SCOPE",
              "RESOURCE_NOT_FOUND",
              "CONTENT_TYPE_UNSUPPORTED",
              "REQUEST_BODY_TOO_LARGE",
              "VALIDATION_FAILED",
              "IDEMPOTENCY_KEY_REQUIRED",
              "IDEMPOTENCY_KEY_REUSED",
              "RATE_LIMITED",
              "INTERNAL_ERROR",
              "DEPENDENCY_UNAVAILABLE",
              "POINT_PACKAGE_AUCTION_INELIGIBLE",
              "LINK_ATTEMPT_ALREADY_FINALIZED",
              "ACTIVE_RESERVATION_EXISTS",
              "INSUFFICIENT_BALANCE",
              "POINT_RESERVATION_NOT_ACTIVE",
              "POINT_RESERVATION_EXPIRED",
              "RESERVATION_VECTOR_HASH_MISMATCH",
              "SETTLEMENT_PLAN_HASH_MISMATCH"
            ]
          },
          "requestId": {
            "$ref": "#/components/schemas/OpaqueId"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "minItems": 1
          }
        },
        "required": ["type", "title", "status", "code", "requestId"]
      },
      "RateLimitedProblem": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string",
            "minLength": 1
          },
          "status": {
            "type": "integer",
            "const": 429
          },
          "detail": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000
          },
          "instance": {
            "type": "string",
            "format": "uri-reference"
          },
          "code": {
            "type": "string",
            "const": "RATE_LIMITED"
          },
          "requestId": {
            "$ref": "#/components/schemas/OpaqueId"
          }
        },
        "required": ["type", "title", "status", "code", "requestId"]
      },
...
      "ReservationStatusByIdRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "lookupBy": {
            "type": "string",
            "const": "POINT_RESERVATION_ID"
          },
          "pointReservationIds": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OpaqueId"
            },
            "minItems": 1,
            "uniqueItems": true
          }
        },
        "required": ["lookupBy", "pointReservationIds"]
      },
      "ReservationStatusByKeyRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "lookupBy": {
            "type": "string",
            "const": "RESERVATION_KEY"
          },
          "reservationKeys": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReservationKey"
            },
            "minItems": 1,
            "uniqueItems": true
          }
        },
        "required": ["lookupBy", "reservationKeys"]
      },
      "ReservationStatusRequest": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/ReservationStatusByIdRequest"
          },
          {
            "$ref": "#/components/schemas/ReservationStatusByKeyRequest"
          }
        ]
      },
      "ReservationStatusItem": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/ActiveReservationStatusItem"
          },
          {
            "$ref": "#/components/schemas/CapturedReservationStatusItem"
          },
          {
            "$ref": "#/components/schemas/ReleasedReservationStatusItem"
          },
          {
            "$ref": "#/components/schemas/ExpiredReservationStatusItem"
          }
        ]
      },
      "ActiveReservationStatusItem": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "pointReservationId": {
            "$ref": "#/components/schemas/OpaqueId"
          },
          "reservationKey": {
            "$ref": "#/components/schemas/ReservationKey"
          },
          "status": {
            "type": "string",
            "const": "ACTIVE"
          },
          "auctionId": {
            "$ref": "#/components/schemas/OpaqueId"
          },
          "settlementId": {
            "$ref": "#/components/schemas/OpaqueId"
          },
          "planHash": {
            "$ref": "#/components/schemas/Sha256Hash"
          },
          "vectorHash": {
            "$ref": "#/components/schemas/Sha256Hash"
          },
          "createdAt": {
            "$ref": "#/components/schemas/UtcInstant"
          },
          "expiresAt": {
            "$ref": "#/components/schemas/UtcInstant"
          },
          "terminalAt": {
            "type": "null"
          },
          "terminalReceiptId": {
            "type": "null"
          }
        },
        "required": [
          "pointReservationId",
          "reservationKey",
          "status",
          "auctionId",
          "settlementId",
          "planHash",
          "vectorHash",
          "createdAt",
          "expiresAt",
          "terminalAt",
          "terminalReceiptId"
        ]
      },
      "CapturedReservationStatusItem": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "pointReservationId": {
            "$ref": "#/components/schemas/OpaqueId"
          },
          "reservationKey": {
            "$ref": "#/components/schemas/ReservationKey"
          },
          "status": {
            "type": "string",
            "const": "CAPTURED"
          },
          "auctionId": {
            "$ref": "#/components/schemas/OpaqueId"
          },
          "settlementId": {
            "$ref": "#/components/schemas/OpaqueId"
          },
          "planHash": {
            "$ref": "#/components/schemas/Sha256Hash"
          },
          "vectorHash": {
            "$ref": "#/components/schemas/Sha256Hash"
          },
          "createdAt": {
            "$ref": "#/components/schemas/UtcInstant"
          },
          "expiresAt": {
            "$ref": "#/components/schemas/UtcInstant"
          },
          "terminalAt": {
            "$ref": "#/components/schemas/UtcInstant"
          },
          "terminalReceiptId": {
            "$ref": "#/components/schemas/OpaqueId"
          }
        },
        "required": [
          "pointReservationId",
          "reservationKey",
          "status",
          "auctionId",
          "settlementId",
          "planHash",
          "vectorHash",
          "createdAt",
          "expiresAt",
          "terminalAt",
          "terminalReceiptId"
        ]
      },
      "ReleasedReservationStatusItem": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "pointReservationId": {
            "$ref": "#/components/schemas/OpaqueId"
          },
          "reservationKey": {
            "$ref": "#/components/schemas/ReservationKey"
          },
          "status": {
            "type": "string",
            "const": "RELEASED"
          },
          "auctionId": {
            "$ref": "#/components/schemas/OpaqueId"
          },
          "settlementId": {
            "$ref": "#/components/schemas/OpaqueId"
          },
          "planHash": {
            "$ref": "#/components/schemas/Sha256Hash"
          },
          "vectorHash": {
            "$ref": "#/components/schemas/Sha256Hash"
          },
          "createdAt": {
            "$ref": "#/components/schemas/UtcInstant"
          },
          "expiresAt": {
            "$ref": "#/components/schemas/UtcInstant"
          },
          "terminalAt": {
            "$ref": "#/components/schemas/UtcInstant"
          },
          "terminalReceiptId": {
            "$ref": "#/components/schemas/OpaqueId"
          }
        },
        "required": [
          "pointReservationId",
          "reservationKey",
          "status",
          "auctionId",
          "settlementId",
          "planHash",
          "vectorHash",
          "createdAt",
          "expiresAt",
          "terminalAt",
          "terminalReceiptId"
        ]
      },
      "ExpiredReservationStatusItem": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "pointReservationId": {
            "$ref": "#/components/schemas/OpaqueId"
          },
          "reservationKey": {
            "$ref": "#/components/schemas/ReservationKey"
          },
          "status": {
            "type": "string",
            "const": "EXPIRED"
          },
          "auctionId": {
            "$ref": "#/components/schemas/OpaqueId"
          },
          "settlementId": {
            "$ref": "#/components/schemas/OpaqueId"
          },
          "planHash": {
            "$ref": "#/components/schemas/Sha256Hash"
          },
          "vectorHash": {
            "$ref": "#/components/schemas/Sha256Hash"
          },
          "createdAt": {
            "$ref": "#/components/schemas/UtcInstant"
          },
          "expiresAt": {
            "$ref": "#/components/schemas/UtcInstant"
          },
          "terminalAt": {
            "$ref": "#/components/schemas/UtcInstant"
          },
          "terminalReceiptId": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/OpaqueId"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "pointReservationId",
          "reservationKey",
          "status",
          "auctionId",
          "settlementId",
          "planHash",
          "vectorHash",
          "createdAt",
          "expiresAt",
          "terminalAt",
          "terminalReceiptId"
        ]
      },
      "ReservationStatusData": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReservationStatusItem"
            }
          }
        },
        "required": ["items"]
      },
      "ReservationStatusResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ReservationStatusData"
          },
          "meta": {
            "$ref": "#/components/schemas/RequestMeta"
          }
        },
        "required": ["data", "meta"]
      },
...
      "ReleaseReservationRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "pointReservationId": {
            "$ref": "#/components/schemas/OpaqueId"
          },
          "reason": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1000
          },
          "planHash": {
            "$ref": "#/components/schemas/Sha256Hash"
          }
        },
        "required": ["pointReservationId", "reason", "planHash"]
      },
      "ReleaseReservationData": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "releaseReceiptId": {
            "$ref": "#/components/schemas/OpaqueId"
          },
          "pointReservationId": {
            "$ref": "#/components/schemas/OpaqueId"
          },
          "status": {
            "type": "string",
            "const": "RELEASED"
          },
          "reason": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1000
          },
          "planHash": {
            "$ref": "#/components/schemas/Sha256Hash"
          },
          "releasedAt": {
            "$ref": "#/components/schemas/UtcInstant"
          },
          "contentHash": {
            "$ref": "#/components/schemas/Sha256Hash"
          }
        },
        "required": [
          "releaseReceiptId",
          "pointReservationId",
          "status",
          "reason",
          "planHash",
          "releasedAt",
          "contentHash"
        ]
      },
      "ReleaseReservationResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ReleaseReservationData"
          },
          "meta": {
            "$ref": "#/components/schemas/RequestMeta"
          }
        },
        "required": ["data", "meta"]
      }
    }
  }
}